+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Summary creation lock should be fixed
+
 ## [0.20.0] - 2022-05-03
 
 ### Added